Adding totalLevelCheck until
