+++ src/utils/Index.js
@@ -1,3 +1,4 @@
 export { default as checkKc } from './checkKc.js';
 export { default as grumblerCheck } from './grumblerCheck.js';
 export { default as sanitizedTime } from './sanitizedTime.js';
+export { default as totalLevelCheck } from './totalLevelCheck.js';
